Updated plotter arm, and fixed mechatronics arm to use chains
